Backend/yelpBackend/src/test/java/com/ex/yelpBackend/controller/UserControllerTest.java
package com.ex.yelpBackend.controller;

import com.ex.yelpBackend.model.User;
import org.junit.Before;
import org.junit.Test;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MvcResult;
import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;

import static org.junit.Assert.*;

public class UserControllerTest extends AbstractTest {

    @Override
    @Before
    public void setUp() {
        super.setUp();
    }

    @Test
    public void getAllUsers() throws Exception{
        String uri = "/yelp/user/all";
        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();

        int status = mvcResult.getResponse().getStatus();
        assertEquals(200, status);
        String content = mvcResult.getResponse().getContentAsString();
        User[] userlist = super.mapFromJson(content, User[].class);
        assertTrue(userlist.length > 0);
    }

    @Test
    public void findByUsername() throws Exception{
        String uri = "/yelp/user/Cody";
        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.get(uri)
                .accept(MediaType.APPLICATION_JSON_VALUE)).andReturn();

        int status = mvcResult.getResponse().getStatus();
        assertEquals(200, status);
        String content = mvcResult.getResponse().getContentAsString();
        User u = super.mapFromJson(content, User.class);
        assertNotNull(u);
    }

    @Test
    public void insertU() throws Exception{
        String uri = "/yelp/user/insert";
        User user = new User();
        user.setName("testtest");
        user.setUsername("testU");
        user.setPassword("1111");
        String inputJson = super.mapToJson(user);
        MvcResult mvcResult = mvc.perform(MockMvcRequestBuilders.post(uri)
                .contentType(MediaType.APPLICATION_JSON_VALUE)
                .content(inputJson)).andReturn();

        int status = mvcResult.getResponse().getStatus();
        assertEquals(201, status);
        String content = mvcResult.getResponse().getContentAsString();
        assertNotNull(content);

    }

    @Test
    public void throwsExceptionOnNull() {
        UserController uc = new UserController();
        uc.insertU(null);
    }

}